+++ src/PackageSystem/Instructions/FileSystem/ApplicationFileSystemInstruction.ts
@@ -71,6 +71,14 @@ export class ApplicationFileSystemInstruction extends FileSystemInstruction
     }
 
     /**
+     * Gets the name of the directory to save the assets archive to.
+     */
+    protected get AssetDirectoryName(): string
+    {
+        return "files";
+    }
+
+    /**
      * @inheritdoc
      *
      * @param source
@@ -81,6 +89,20 @@ export class ApplicationFileSystemInstruction extends FileSystemInstruction
      */
     protected override MakeDefaultFileName(source: string): string
     {
-        return join(...(this.Application ? [this.Application] : []), `${super.MakeDefaultFileName(source)}.tar`);
+        return join(this.AssetDirectoryName, ...(this.Application ? [this.Application] : []), `${this.GetAssetFileName(source)}.tar`);
+    }
+
+    /**
+     * Gets the default name of the asset file.
+     *
+     * @param source
+     * The source of the instruction.
+     *
+     * @returns
+     * The default name of the asset file.
+     */
+    protected GetAssetFileName(source: string): string
+    {
+        return super.MakeDefaultFileName(source);
     }
 }